pyproject.toml
[tool.poetry]
name = "x-request-id-middleware"
version = "0.2.1"
description = "Library to handle request ID propagation for Django and FastAPI"
authors = ["WatchingWhileUSleep <whatchingwhileusleepx@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/watchingwhileusleep/x-request-id-middleware"
repository = "https://github.com/watchingwhileusleep/x-request-id-middleware"
documentation = "https://github.com/watchingwhileusleep/x-request-id-middleware?tab=readme-ov-file"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
sentry-sdk = "^1.16.0"

[tool.poetry.extras]
django = ["django^4.2"]
fastapi = ["fastapi^0.103", "starlette^0.27", "anyio^3.6.2", "httpx^0.27.0"]

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5"
pytest-cov = "^4.0"
factory-boy = "^3.2"
mypy = "^1.4"
ruff = "^0.6.2"
pytest = "^7.4"

[tool.ruff]
line-length = 80
target-version = "py38"

[tool.ruff.lint]
    ignore = [
        "DJ008",
        "PT001",
        "PT023",
        "SIM103",
    ]
    select = [
        "B",    # flake8-bugbear,
        "C4",   # flake8-comprehensions
        "DJ",   # flake8-django
        "E",    # pycodestyle error
        "F",    # pyflakes
        "SIM",  # flake8-simplify
        "T20",  # flake8-print
        "W",    # pycodestyle warning
        "UP",   # pyupgrade
        "I",    # isort
    ]

[tool.isort]
line_length = 80
multi_line_output = 3
force_single_line = true
include_trailing_comma = true
sections = ["STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
